requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
discord.py==1.6.0
Faker==7.0.1
faker-vehicle==0.2.0
idna==3.1
multidict==5.1.0
pycparser==2.20
pyfiglet==0.8.post1
python-dateutil==2.8.1
RandomWords==0.3.0
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
yarl==1.6.3
